// model/cardRuleNumberCondition.ts
/**
 * Gr4vy API (Beta)
 * Welcome to the Gr4vy API reference documentation. Our API is still very much a work in product and subject to change.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@gr4vy.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models'

/**
 * Part of a rule that matches numeric fields. It defines the condition under which this rule applies.
 */
export class CardRuleNumberCondition {
  /**
   * `number`.
   */
  'match': CardRuleNumberCondition.MatchEnum
  /**
   * The transaction field to filter by.
   */
  'key': CardRuleNumberCondition.KeyEnum
  /**
   * The comparison to make to `value` property.
   */
  'operator': CardRuleNumberCondition.OperatorEnum
  /**
   * The values to compare the `key` to.
   */
  'value': number

  static discriminator: string | undefined = undefined

  static attributeTypeMap: Array<{
    name: string
    baseName: string
    type: string
  }> = [
    {
      name: 'match',
      baseName: 'match',
      type: 'CardRuleNumberCondition.MatchEnum',
    },
    {
      name: 'key',
      baseName: 'key',
      type: 'CardRuleNumberCondition.KeyEnum',
    },
    {
      name: 'operator',
      baseName: 'operator',
      type: 'CardRuleNumberCondition.OperatorEnum',
    },
    {
      name: 'value',
      baseName: 'value',
      type: 'number',
    },
  ]

  static getAttributeTypeMap() {
    return CardRuleNumberCondition.attributeTypeMap
  }
}

export namespace CardRuleNumberCondition {
  export enum MatchEnum {
    Number = <any>'number',
  }
  export enum KeyEnum {
    Amount = <any>'amount',
  }
  export enum OperatorEnum {
    Equal = <any>'==',
    NotEqual = <any>'!=',
    LessThanOrEqualTo = <any>'<=',
    GreaterThanOrEqualTo = <any>'>=',
    GreaterThan = <any>'>',
    LessThan = <any>'<',
  }
}