// model/error404PendingCreation.ts
/**
 * Gr4vy API
 * Welcome to the Gr4vy API reference documentation. Our API is still very much a work in product and subject to change.
 *
 * The version of the OpenAPI document: 1.1.0-beta
 * Contact: code@gr4vy.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ErrorDetail } from './errorDetail';

/**
* Pending Creation Error (HTTP 404).
*/
export class Error404PendingCreation {
    /**
    * `error`.
    */
    'type'?: Error404PendingCreation.TypeEnum;
    /**
    * `pending_creation`.
    */
    'code'?: Error404PendingCreation.CodeEnum;
    /**
    * `404`.
    */
    'status'?: Error404PendingCreation.StatusEnum;
    /**
    * The resource is still pending.
    */
    'message'?: Error404PendingCreation.MessageEnum;
    /**
    * A list of detail objects that further clarify the reason for the error. Not every error supports more detail.
    */
    'details'?: Array<ErrorDetail>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "Error404PendingCreation.TypeEnum"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "Error404PendingCreation.CodeEnum"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "Error404PendingCreation.StatusEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "Error404PendingCreation.MessageEnum"
        },
        {
            "name": "details",
            "baseName": "details",
            "type": "Array<ErrorDetail>"
        }    ];

    static getAttributeTypeMap() {
        return Error404PendingCreation.attributeTypeMap;
    }
}

export namespace Error404PendingCreation {
    export enum TypeEnum {
        Error = <any> 'error'
    }
    export enum CodeEnum {
        PendingCreation = <any> 'pending_creation'
    }
    export enum StatusEnum {
        NUMBER_404 = <any> 404
    }
    export enum MessageEnum {
        TheResourceIsStillPending = <any> 'The resource is still pending'
    }
}